toy program ongoing

+++ src/ToonEngine/ToonEngine/ToonCRC32Hash.cpp
@@ -12,7 +12,7 @@ namespace Toon
 	code or tables extracted from it, as desired without restriction.*/
 	unsigned long CRC32Hasher::hashRuntime(char const * buf, unsigned long long len)
 	{
-		DWORD oldcrc32 = 0xFFFFFFFF;
+		unsigned long oldcrc32 = 0xFFFFFFFF;
 		for (; len; --len, ++buf)
 		{
 			oldcrc32 = UPDC32(*buf, oldcrc32);
